Attempt to fix the Keep Save Files option.
Soemtimes the save file gets deleted when I die even if I have the
Keep Save Files option checked.  I haven't figured out why that happens,
so change the default to be that deleteGame doesn't delete the file if
that option is set, unless a new "force" option is specified, and only
pass "force = true" in the cases where the file should always be deleted
(mostly when ending the game after getting the amulet).

+++ app/src/main/java/com/watabou/pixeldungeon/Dungeon.java
@@ -552,8 +552,11 @@ public class Dungeon {
 		return (Level)bundle.get( "level" );
 	}
 	
-	public static void deleteGame( HeroClass cl, boolean deleteLevels ) {
-		
+	public static void deleteGame( HeroClass cl, boolean deleteLevels, boolean force ) {
+
+		if (!force && PixelDungeon.keepSaveFiles())
+			return;
+
 		Game.instance.deleteFile( gameFile( cl ) );
 		
 		if (deleteLevels) {
@@ -565,7 +568,11 @@ public class Dungeon {
 		
 		GamesInProgress.delete( cl );
 	}
-	
+
+	public static void deleteGame( HeroClass cl, boolean deleteLevels ) {
+		deleteGame(cl, deleteLevels, false);
+	}
+
 	public static Bundle gameBundle( String fileName ) throws IOException {
 		
 		InputStream input = Game.instance.openFileInput( fileName );
